refactor(lexer): extract token definitions into separate module
- Created src/lexer/ directory structure
- Moved Token enum (52 variants) to src/lexer/token.rs
- Moved is_keyword() and is_shell_keyword() functions to token.rs
- Created src/lexer/mod.rs with all lexing logic
- Removed src/lexer.rs (replaced by module structure)
- Added comprehensive module-level documentation
- All 1,260 tests pass
- Maintains backward compatibility through re-exports

This is Step 3.1 of Phase 3 lexer refactoring. The token definitions
(~200 lines) are now cleanly separated from the lexing logic (~3,400 lines),
improving code organization and maintainability.
